requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.2
river==0.21.0
python-dotenv==1.0.1
gunicorn==21.2.0